fix freeze bug
Here I just simply added an if-statement to check if the pause menu was toggled.
This can set the timescale of the game accordingly.

The previous issue without doing this, was that when the player unpaused with the escape key, the game was still in a frozen state, as the timescale hasn't been reset to 1f.